+++ module/contextmenus.js
@@ -7,10 +7,16 @@ import { ConfigSheetForActions } from "./ConfigSheetForActions.js";
  */
 export const onGetCardsDirectoryEntryContext = (html, options) => {
 
-    const hiddenActionNames = ["FOLDER.Clear", "SIDEBAR.Delete", "SIDEBAR.Duplicate", "PERMISSION.Configure", "SIDEBAR.Export", "SIDEBAR.Import"];
     // Classic entries are only available for decks with no rtucards flag
+    let hiddenActionNames = ["FOLDER.Clear", "SIDEBAR.Delete", "SIDEBAR.Duplicate", "SIDEBAR.Export", "SIDEBAR.Import"];
     options.filter( o => hiddenActionNames.includes(o.name) ).forEach( option => {
-        option.condition = appendRTUCardsFlagCondition(option.condition, false);
+        option.condition = appendRTUCardsFlagCondition(option.condition, {forRegisteredStacks: false} );
+    });
+
+    // Permission control is only for decks and discards
+    hiddenActionNames = ["PERMISSION.Configure"];
+    options.filter( o => hiddenActionNames.includes(o.name) ).forEach( option => {
+        option.condition = appendRTUCardsFlagCondition(option.condition, {forRegisteredStacks: false, decksAndDiscardAllowed: true} );
     });
 
     addConfigureActionsEntry(options);
@@ -18,14 +24,21 @@ export const onGetCardsDirectoryEntryContext = (html, options) => {
     addUnregisterEntry(options);
 }
 
-const appendRTUCardsFlagCondition = (currentCondition, flagShouldBeHere) => {
+const appendRTUCardsFlagCondition = (currentCondition, {forRegisteredStacks=false, decksAndDiscardAllowed=false}={}) => {
     const newCondition = (li) => {
         if( currentCondition && !currentCondition(li) ) { return false; }
 
         const stack = game.cards.get(li.data("documentId"));
         const custom = new CustomCardStack(stack);
-        if( flagShouldBeHere ) { return custom.handledByModule; }
-        return !custom.handledByModule;
+
+        const handled = custom.handledByModule;
+        if(handled) {
+            if( decksAndDiscardAllowed ) {
+                return custom.stackOwner.forNobody;
+            }
+            return forRegisteredStacks;
+        }
+        return !forRegisteredStacks;
     }
     return newCondition;
 }
@@ -54,7 +67,7 @@ const addConfigureActionsEntry = (allEntries) => {
             sheet.render(true);
         }
     };
-    entry.condition = appendRTUCardsFlagCondition(entry.condition, true);
+    entry.condition = appendRTUCardsFlagCondition(entry.condition, {forRegisteredStacks: true} );
     allEntries.push(entry);
 }
 
@@ -74,7 +87,7 @@ const addRegisterEntry = (allEntries) => {
             await custom.registerAsHandledByModule();
         }
     };
-    entry.condition = appendRTUCardsFlagCondition(entry.condition, false);
+    entry.condition = appendRTUCardsFlagCondition(entry.condition, {forRegisteredStacks: false});
     allEntries.push(entry);
 }
 
@@ -95,6 +108,6 @@ const addUnregisterEntry = (allEntries) => {
             await custom.unregisterAsHandledByModule();
         }
     };
-    entry.condition = appendRTUCardsFlagCondition(entry.condition, true);
+    entry.condition = appendRTUCardsFlagCondition(entry.condition, {forRegisteredStacks: false});
     allEntries.push(entry);
 }